+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.1] 2024-02-19
+
+### Removed
+
+- `TestCampaign.plot_multipactor_limits`
+- `MultipactorTest.plot_multipactor_limits`
+- `MultipactorTest._get_proper_multipactor_bands`
+- `MultipactorTest.filter_measurement_points`
+- In following methods, `multipactor_measured_at` is no longer supported. `MultipactorBands` object(s) must be given instead.
+  - `MultipactorTest.data_for_somersalo`
+  - `MultipactorTest.data_for_susceptibility`
+- `Instruments.RfPower`
+- `Instruments.SWR`
+
+
 ## [1.4.0] 2024-02-14
 
 ### Added